pyproject.toml
[project]
name = "serious"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyaudio>=0.2.14",
    "pyautogui>=0.9.54 ; sys_platform == 'linux'",
    "torch>=2.7.0",
    "transformers>=4.52.4",
    "pyobjc-framework-Quartz ; sys_platform == 'darwin'"
]
